refactor: rename shrinkray.subprocess to shrinkray.interp
Mechanical rename ahead of moving the TUI/reducer split from a worker
subprocess to a subinterpreter in the same process. No behaviour
change.
