+++ src/shrinkray/__main__.py
@@ -468,7 +468,7 @@ def main(
 def worker_main() -> None:
     """Entry point for the worker subprocess."""
     # Lazy import to avoid loading worker module in main process (fast CLI startup)
-    from shrinkray.subprocess.worker import (  # noqa: I001, no-import-in-function
+    from shrinkray.interp.worker import (  # noqa: I001, no-import-in-function
         main as worker_entry,
     )
