+++ src/shrinkray/interp/protocol.py
