--- src/shrinkray/subprocess/__init__.py
@@ -1,20 +1,0 @@
-"""Subprocess communication for separating reducer from UI."""
-
-from shrinkray.subprocess.client import SubprocessClient
-from shrinkray.subprocess.protocol import (
-    ProgressUpdate,
-    Request,
-    Response,
-    deserialize,
-    serialize,
-)
-
-
-__all__ = [
-    "Request",
-    "Response",
-    "ProgressUpdate",
-    "serialize",
-    "deserialize",
-    "SubprocessClient",
-]
